test(person360): the last-touch fold gets its own file
The relationship room's suite is about what a caller is REFUSED. Attribution
is a different question and needs participant fixtures no refusal test has a
use for, so it moves out rather than pushing that file past its ceiling.
